+++ packages/flow-core/src/components/f-button/f-button.ts
@@ -110,6 +110,7 @@ export class FButton extends FElement {
 
   render() {
     this.validateProperties();
+    this.innerHTML = "";
 
     /**
      * create iconLeft if available