+++ packages/flow-core/src/components/f-counter/f-counter.ts
@@ -52,6 +52,7 @@ export class FCounter extends FElement {
     }
   }
   render() {
+    this.innerHTML = "";
     this.validateProperties();
     return html`${this.loading
       ? html`${unsafeSVG(loader)}`