+++ packages/flow-core/src/components/f-icon/f-icon.ts
@@ -106,6 +106,7 @@ export class FIcon extends FElement {
   }
 
   render() {
+    this.innerHTML = "";
     this.validateProperties();
     return html`${this.loading
       ? html`${unsafeSVG(loader)}`